fix: Plan → plan 폴더명 케이싱 수정
