+++ components/features/plan/LobbyPlanItem.tsx
